Fixing up FEATURE_EVENTSOURCE_XPLAT
