+++ src/vm/eventtrace.cpp
@@ -7412,31 +7412,4 @@ VOID ETW::EnumerationLog::EnumerationHelper(Module *moduleFilter, BaseDomain *do
     }    
 }
 
-#if defined(FEATURE_EVENTSOURCE_XPLAT)
-
-void QCALLTYPE XplatEventSourceLogger::LogEventSource(__in_z int eventID, __in_z LPCWSTR eventName, __in_z LPCWSTR eventSourceName, __in_z LPCWSTR payload)
-{
-    QCALL_CONTRACT;
-
-    BEGIN_QCALL;
-    FireEtwEventSource(eventID, eventName, eventSourceName, payload);
-    END_QCALL;
-}
-
-BOOL QCALLTYPE XplatEventSourceLogger::IsEventSourceLoggingEnabled()
-{
-    QCALL_CONTRACT;
-
-    BOOL retVal = FALSE;
-
-    BEGIN_QCALL;
-    retVal = XplatEventLogger::IsEventLoggingEnabled();
-    END_QCALL;
-    
-    return retVal;
-
-}
-
-#endif //defined(FEATURE_EVENTSOURCE_XPLAT)
-
 #endif // !FEATURE_REDHAWK
