+++ src/vm/eventtracepriv.h
@@ -408,17 +408,3 @@ private:
 
 #endif // __EVENTTRACEPRIV_H__
 
-#if defined(FEATURE_EVENTSOURCE_XPLAT)
-class XplatEventSourceLogger
-{
-public:
-#ifdef FEATURE_EVENT_TRACE
-    static void QCALLTYPE LogEventSource(__in_z int eventID, __in_z LPCWSTR eventName, __in_z LPCWSTR eventSourceName, __in_z LPCWSTR payload);
-    static BOOL QCALLTYPE IsEventSourceLoggingEnabled();
-#else
-    static void QCALLTYPE LogEventSource(__in_z int eventID, __in_z LPCWSTR eventName, __in_z LPCWSTR eventSourceName, __in_z LPCWSTR payload) {}
-    static BOOL QCALLTYPE IsEventSourceLoggingEnabled() { return false; }
-#endif
-};
-
-#endif //defined(FEATURE_EVENTSOURCE_XPLAT)
