+++ src/vm/mscorlib.cpp
@@ -131,10 +131,7 @@
 #endif
 
 #if defined(FEATURE_EVENTSOURCE_XPLAT)
-
-#define __EVENTTRACEPRIV_H__
-#include "eventtracepriv.h"
-#undef __EVENTTRACEPRIV_H__
+#include "nativeeventsource.h"
 #endif //defined(FEATURE_EVENTSOURCE_XPLAT)
 
 #endif // CROSSGEN_MSCORLIB
