+++ src/vm/nativeeventsource.cpp
@@ -1,0 +1,38 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+//
+// File: nativeeventsource.cpp
+// Abstract: This module implements native part of Event Source support in VM
+//
+//
+//
+// ============================================================================
+
+#include "common.h"
+#if defined(FEATURE_EVENTSOURCE_XPLAT)
+#include "nativeeventsource.h"
+
+void QCALLTYPE XplatEventSourceLogger::LogEventSource(__in_z int eventID, __in_z LPCWSTR eventName, __in_z LPCWSTR eventSourceName, __in_z LPCWSTR payload)
+{
+    QCALL_CONTRACT;
+    BEGIN_QCALL;
+    FireEtwEventSource(eventID, eventName, eventSourceName, payload);
+    END_QCALL;
+}
+
+BOOL QCALLTYPE XplatEventSourceLogger::IsEventSourceLoggingEnabled()
+{
+    QCALL_CONTRACT;
+
+    BOOL retVal = FALSE;
+
+    BEGIN_QCALL;
+    retVal = XplatEventLogger::IsEventLoggingEnabled();
+    END_QCALL;
+    
+    return retVal;
+
+}
+
+#endif //defined(FEATURE_EVENTSOURCE_XPLAT)
